refactor: Remove duplicate code from Python test files and move it into conftest.py
