+++ tests/test_find_unreleased_changelog_items.py
@@ -77,19 +77,6 @@ def fixture_summary_file(tmp_path: Path) -> Path:
     return tmp_path / "github_summary.txt"
 
 
-@pytest.fixture(name="mock_github_output_file")
-def fixture_mock_github_output_file(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> Path:
-    """Mock the GitHub output file.
-
-    Args:
-        tmp_path (fixture): The temporary path fixture.
-        monkeypatch (fixture): The monkeypatch fixture.
-    """
-    github_output_file = tmp_path / "github_output"
-    monkeypatch.setenv("GITHUB_OUTPUT", str(github_output_file))
-    return github_output_file
-
-
 @pytest.fixture(name="mock_env_vars")
 def fixture_mock_env_vars(
     tmp_path: Path,
